Cleans Up home_fragment.xml and HomeFragment.java
Now it uses a nice ConstraintLayout with Sliders for numeric input.
There are further improvements to be made like theming the colours of
the sliders/inputs, displaying help info, making a nicer background
colour, bits like that.

This may have introduced a bug in MainService.onDestroy where the
locationManager is null, but I don't see how this PR could have made
that crop up. Either way, will tune that up next time I'm working on
this.